Type Alias core_strategy::StrategyConfig

pub type StrategyConfig = MultiStrategyConfig;

Aliased Type

struct StrategyConfig {
    pub on_fail_continue: bool,
    pub allow_recursive: bool,
    pub finalize_channel_closure: bool,
    pub strategies: Vec<Strategy>,
}

Fields
on_fail_continue: bool

Determines if the strategy should continue executing the next strategy if the current one failed. If set to true, the strategy behaves like a logical AND chain of SingularStrategies Otherwise, it behaves like a logical OR chain of SingularStrategies. Default is true.

allow_recursive: bool

Indicate whether the MultiStrategy can contain another MultiStrategy. Default is true.

finalize_channel_closure: bool

Indicates if the strategy should check for PendingToClose channels which have elapsed the closure grace period, to issue another channel closing transaction to close them. If not set, the user has to trigger the channel closure manually once again after the grace period is over. Default: false

strategies: Vec<Strategy>

Configuration of individual sub-strategies. Default is empty, which makes the MultiStrategy behave as passive.